getFloat to getLong

+++ src/org/jgtdsl/reports/GasConsumptionReport.java
@@ -1822,9 +1822,9 @@ public class GasConsumptionReport extends BaseAction {
         		meterReadingDto.setActual_consumption(resultSet.getFloat("ACTUAL_CONSUMPTION"));
         		meterReadingDto.setPressure(resultSet.getFloat("PRESSURE"));
         		meterReadingDto.setPressure_factor(resultSet.getFloat("PRESSURE_FACTOR"));
-        		meterReadingDto.setPrev_reading(resultSet.getFloat("PREV_READING"));
-        		meterReadingDto.setCurr_reading(resultSet.getFloat("CURR_READING"));
-        		meterReadingDto.setDifference(resultSet.getFloat("DIFFERENCE"));
+        		meterReadingDto.setPrev_reading(resultSet.getLong("PREV_READING"));
+        		meterReadingDto.setCurr_reading(resultSet.getLong("CURR_READING"));
+        		meterReadingDto.setDifference(resultSet.getLong("DIFFERENCE"));
         		meterReadingDto.setMax_load(resultSet.getFloat("PMAX_LOAD"));
         		meterReadingDto.setMin_load(resultSet.getFloat("PMIN_LOAD"));
         		meterReadingDto.setMeter_rent(resultSet.getFloat("METER_RENT"));
